Move devices module to services package to avoid circular import with mp spawn method
