+++ src/metro/devices/abstract/parallel_operator.py
@@ -25,9 +25,7 @@ import threading
 import traceback
 
 import metro
-# For some strange reason we need to import this explicitely for the
-# second process when using the spawn method for new processes.
-from metro.devices import WidgetDevice
+metro.init_mp_support()
 
 
 _targets = {}
@@ -187,7 +185,7 @@ class Agent(metro.QObject):
         self.stepDone.emit()
 
 
-class Device(WidgetDevice):
+class Device(metro.WidgetDevice):
     def prepare(self, operator_cls, operator_args, newData, state,
                 prefilter=lambda d: d, target=None, target_cap=5,
                 data_pipes=None):
