+++ src/metro/services/devices.py
